Changes related to multiple

+++ OPDCLAIMFORM/Controllers/HOSPITALEXPENSEController.cs
@@ -25,54 +25,77 @@ namespace OPDCLAIMFORM.Controllers
         {
             MedicalInfoEntities entities = new MedicalInfoEntities();
 
-            if (id == null)
+            if (Request.IsAuthenticated)
             {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                AuthenticateUser();
+
+                if (id == null)
+                {
+                    return RedirectToAction("Index", "OPDEXPENSEs");
+                }
+
+                OPDEXPENSE oPDEXPENSE = db.OPDEXPENSEs.Find(id);
+
+
+                var result2 = new HOSPITALEXPENSE_MASTERDETAIL()
+                {
+                    listOPDEXPENSEPATIENT = entities.OPDEXPENSE_PATIENT.Where(e => e.OPDEXPENSE_ID == id).ToList(),
+                    listOPDEXPENSEIMAGE = entities.OPDEXPENSE_IMAGE.Where(e => e.OPDEXPENSE_ID == id).ToList(),
+                    OPDEXPENSE_ID = oPDEXPENSE.OPDEXPENSE_ID,
+                    CLAIMANT_SUFFERED_ILLNESS = oPDEXPENSE.CLAIMANT_SUFFERED_ILLNESS,
+                    CLAIMANT_SUFFERED_ILLNESS_DETAILS = oPDEXPENSE.CLAIMANT_SUFFERED_ILLNESS_DETAILS,
+                    CLAIMANT_SUFFERED_ILLNESS_DATE = oPDEXPENSE.CLAIMANT_SUFFERED_ILLNESS_DATE,
+                    DATE_ILLNESS_NOTICED = oPDEXPENSE.DATE_ILLNESS_NOTICED,
+                    DATE_RECOVERY = oPDEXPENSE.DATE_RECOVERY,
+                    DIAGNOSIS = oPDEXPENSE.DIAGNOSIS,
+                    DOCTOR_NAME = oPDEXPENSE.DOCTOR_NAME,
+                    DRUGS_PRESCRIBED_BOOL = oPDEXPENSE.DRUGS_PRESCRIBED_BOOL,
+                    DRUGS_PRESCRIBED_DESCRIPTION = oPDEXPENSE.DRUGS_PRESCRIBED_DESCRIPTION,
+                    EMPLOYEE_DEPARTMENT = oPDEXPENSE.EMPLOYEE_DEPARTMENT,
+                    EMPLOYEE_NAME = oPDEXPENSE.EMPLOYEE_NAME,
+                    FINANCE_APPROVAL = oPDEXPENSE.FINANCE_APPROVAL,
+                    FINANCE_COMMENT = oPDEXPENSE.FINANCE_COMMENT,
+                    FINANCE_NAME = oPDEXPENSE.FINANCE_NAME,
+                    HOSPITAL_NAME = oPDEXPENSE.HOSPITAL_NAME,
+                    HR_APPROVAL = oPDEXPENSE.HR_APPROVAL,
+                    HR_COMMENT = oPDEXPENSE.HR_COMMENT,
+                    HR_NAME = oPDEXPENSE.HR_NAME,
+                    MANAGEMENT_APPROVAL = oPDEXPENSE.MANAGEMENT_APPROVAL,
+                    MANAGEMENT_COMMENT = oPDEXPENSE.MANAGEMENT_COMMENT,
+                    MANAGEMENT_NAME = oPDEXPENSE.MANAGEMENT_NAME,
+                    PERIOD_CONFINEMENT_DATE_FROM = oPDEXPENSE.PERIOD_CONFINEMENT_DATE_FROM,
+                    PERIOD_CONFINEMENT_DATE_TO = oPDEXPENSE.PERIOD_CONFINEMENT_DATE_TO,
+                    STATUS = oPDEXPENSE.STATUS,
+                    TOTAL_AMOUNT_CLAIMED = oPDEXPENSE.TOTAL_AMOUNT_CLAIMED,
+                    CLAIM_YEAR = oPDEXPENSE.CLAIM_YEAR,
+                    CREATED_DATE = oPDEXPENSE.CREATED_DATE,
+                    MODIFIED_DATE = oPDEXPENSE.MODIFIED_DATE
+
+
+                };
+
+                return View(result2);
             }
-
-            OPDEXPENSE oPDEXPENSE = db.OPDEXPENSEs.Find(id);
-       
-
-            var result2 = new HOSPITALEXPENSE_MASTERDETAIL()
+            else
             {
-                listOPDEXPENSEPATIENT = entities.OPDEXPENSE_PATIENT.Where(e => e.OPDEXPENSE_ID == id).ToList(),
-                listOPDEXPENSEIMAGE = entities.OPDEXPENSE_IMAGE.Where(e => e.OPDEXPENSE_ID == id).ToList(),
-                OPDEXPENSE_ID = oPDEXPENSE.OPDEXPENSE_ID,
-                CLAIMANT_SUFFERED_ILLNESS = oPDEXPENSE.CLAIMANT_SUFFERED_ILLNESS,
-                CLAIMANT_SUFFERED_ILLNESS_DETAILS = oPDEXPENSE.CLAIMANT_SUFFERED_ILLNESS_DETAILS,
-                CLAIMANT_SUFFERED_ILLNESS_DATE = oPDEXPENSE.CLAIMANT_SUFFERED_ILLNESS_DATE,
-                DATE_ILLNESS_NOTICED = oPDEXPENSE.DATE_ILLNESS_NOTICED,
-                DATE_RECOVERY = oPDEXPENSE.DATE_RECOVERY,
-                DIAGNOSIS = oPDEXPENSE.DIAGNOSIS,
-                DOCTOR_NAME = oPDEXPENSE.DOCTOR_NAME,
-                DRUGS_PRESCRIBED_BOOL = oPDEXPENSE.DRUGS_PRESCRIBED_BOOL,
-                DRUGS_PRESCRIBED_DESCRIPTION = oPDEXPENSE.DRUGS_PRESCRIBED_DESCRIPTION,
-                EMPLOYEE_DEPARTMENT = oPDEXPENSE.EMPLOYEE_DEPARTMENT,
-                EMPLOYEE_NAME = oPDEXPENSE.EMPLOYEE_NAME,
-                FINANCE_APPROVAL = oPDEXPENSE.FINANCE_APPROVAL,
-                FINANCE_COMMENT = oPDEXPENSE.FINANCE_COMMENT,
-                FINANCE_NAME = oPDEXPENSE.FINANCE_NAME,
-                HOSPITAL_NAME = oPDEXPENSE.HOSPITAL_NAME,
-                HR_APPROVAL = oPDEXPENSE.HR_APPROVAL,
-                HR_COMMENT = oPDEXPENSE.HR_COMMENT,
-                HR_NAME = oPDEXPENSE.HR_NAME,
-                MANAGEMENT_APPROVAL = oPDEXPENSE.MANAGEMENT_APPROVAL,
-                MANAGEMENT_COMMENT = oPDEXPENSE.MANAGEMENT_COMMENT,
-                MANAGEMENT_NAME = oPDEXPENSE.MANAGEMENT_NAME,
-                PERIOD_CONFINEMENT_DATE_FROM = oPDEXPENSE.PERIOD_CONFINEMENT_DATE_FROM,
-                PERIOD_CONFINEMENT_DATE_TO = oPDEXPENSE.PERIOD_CONFINEMENT_DATE_TO,
-                STATUS = oPDEXPENSE.STATUS,
-                TOTAL_AMOUNT_CLAIMED = oPDEXPENSE.TOTAL_AMOUNT_CLAIMED,
-                CLAIM_YEAR = oPDEXPENSE.CLAIM_YEAR
-            };
-
-            return View(result2);
+                return RedirectToAction("Index", "OPDEXPENSEs");
+            }
         }
 
         // GET: OPDEXPENSEs/Create
         public ActionResult Create()
         {
-            return View();
+            if (Request.IsAuthenticated)
+            {
+                AuthenticateUser();
+
+                return View();
+            }
+            else
+            {
+                return RedirectToAction("Index", "OPDEXPENSEs");
+            }
+
         }
 
         // POST: OPDEXPENSEs/Create
@@ -91,7 +114,7 @@ namespace OPDCLAIMFORM.Controllers
                 db.OPDEXPENSEs.Add(oPDEXPENSE);
                 db.SaveChanges();
                 //return RedirectToAction("Index");
-                return RedirectToAction("Index", "OPDEXPENSEPATIENT", new { id = oPDEXPENSE.OPDEXPENSE_ID });
+                return RedirectToAction("Edit", "HOSPITALEXPENSE", new { id = oPDEXPENSE.OPDEXPENSE_ID });
             }
 
             return View(oPDEXPENSE);
@@ -100,61 +123,64 @@ namespace OPDCLAIMFORM.Controllers
         // GET: OPDEXPENSEs/Edit/5
         public ActionResult Edit(int? id)
         {
-            if (id == null)
+            if (Request.IsAuthenticated)
             {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                AuthenticateUser();
+
+                if (id == null)
+                {
+                    return RedirectToAction("Index", "OPDEXPENSEs");
+                }
+
+                MedicalInfoEntities entities = new MedicalInfoEntities();
+                OPDEXPENSE oPDEXPENSE = db.OPDEXPENSEs.Find(id);
+
+
+                var result2 = new HOSPITALEXPENSE_MASTERDETAIL()
+                {
+                    listOPDEXPENSEPATIENT = entities.OPDEXPENSE_PATIENT.Where(e => e.OPDEXPENSE_ID == id).ToList(),
+                    listOPDEXPENSEIMAGE = entities.OPDEXPENSE_IMAGE.Where(e => e.OPDEXPENSE_ID == id).ToList(),
+                    OPDEXPENSE_ID = oPDEXPENSE.OPDEXPENSE_ID,
+                    CLAIMANT_SUFFERED_ILLNESS = oPDEXPENSE.CLAIMANT_SUFFERED_ILLNESS,
+                    CLAIMANT_SUFFERED_ILLNESS_DETAILS = oPDEXPENSE.CLAIMANT_SUFFERED_ILLNESS_DETAILS,
+                    CLAIMANT_SUFFERED_ILLNESS_DATE = oPDEXPENSE.CLAIMANT_SUFFERED_ILLNESS_DATE,
+
+
+                    DATE_ILLNESS_NOTICED = oPDEXPENSE.DATE_ILLNESS_NOTICED,
+                    DATE_RECOVERY = oPDEXPENSE.DATE_RECOVERY,
+                    DIAGNOSIS = oPDEXPENSE.DIAGNOSIS,
+                    DOCTOR_NAME = oPDEXPENSE.DOCTOR_NAME,
+                    DRUGS_PRESCRIBED_BOOL = oPDEXPENSE.DRUGS_PRESCRIBED_BOOL,
+                    DRUGS_PRESCRIBED_DESCRIPTION = oPDEXPENSE.DRUGS_PRESCRIBED_DESCRIPTION,
+                    EMPLOYEE_DEPARTMENT = oPDEXPENSE.EMPLOYEE_DEPARTMENT,
+                    EMPLOYEE_NAME = oPDEXPENSE.EMPLOYEE_NAME,
+                    FINANCE_APPROVAL = oPDEXPENSE.FINANCE_APPROVAL,
+                    FINANCE_COMMENT = oPDEXPENSE.FINANCE_COMMENT,
+                    FINANCE_NAME = oPDEXPENSE.FINANCE_NAME,
+                    HOSPITAL_NAME = oPDEXPENSE.HOSPITAL_NAME,
+                    HR_APPROVAL = oPDEXPENSE.HR_APPROVAL,
+                    HR_COMMENT = oPDEXPENSE.HR_COMMENT,
+                    HR_NAME = oPDEXPENSE.HR_NAME,
+                    MANAGEMENT_APPROVAL = oPDEXPENSE.MANAGEMENT_APPROVAL,
+                    MANAGEMENT_COMMENT = oPDEXPENSE.MANAGEMENT_COMMENT,
+                    MANAGEMENT_NAME = oPDEXPENSE.MANAGEMENT_NAME,
+                    PERIOD_CONFINEMENT_DATE_FROM = oPDEXPENSE.PERIOD_CONFINEMENT_DATE_FROM,
+                    PERIOD_CONFINEMENT_DATE_TO = oPDEXPENSE.PERIOD_CONFINEMENT_DATE_TO,
+                    STATUS = oPDEXPENSE.STATUS,
+                    OPDTYPE = oPDEXPENSE.OPDTYPE,
+                    TOTAL_AMOUNT_CLAIMED = oPDEXPENSE.TOTAL_AMOUNT_CLAIMED,
+                    CLAIM_YEAR = oPDEXPENSE.CLAIM_YEAR,
+                    CREATED_DATE = oPDEXPENSE.CREATED_DATE,
+                    MODIFIED_DATE = oPDEXPENSE.MODIFIED_DATE
+                };
+
+                ViewData["OPDEXPENSE_ID"] = id;
+                return View(result2);
             }
-            //OPDEXPENSE oPDEXPENSE = db.OPDEXPENSEs.Find(id);
-            //if (oPDEXPENSE == null)
-            //{
-            //    return HttpNotFound();
-            //}
-            //return View(oPDEXPENSE);
-
-            MedicalInfoEntities entities = new MedicalInfoEntities();
-            OPDEXPENSE oPDEXPENSE = db.OPDEXPENSEs.Find(id);
-
-            
-
-
-            var result2 = new HOSPITALEXPENSE_MASTERDETAIL()
+            else
             {
-                listOPDEXPENSEPATIENT = entities.OPDEXPENSE_PATIENT.Where(e => e.OPDEXPENSE_ID == id).ToList(),
-                listOPDEXPENSEIMAGE = entities.OPDEXPENSE_IMAGE.Where(e => e.OPDEXPENSE_ID == id).ToList(),
-                OPDEXPENSE_ID = oPDEXPENSE.OPDEXPENSE_ID,
-                CLAIMANT_SUFFERED_ILLNESS = oPDEXPENSE.CLAIMANT_SUFFERED_ILLNESS,
-                CLAIMANT_SUFFERED_ILLNESS_DETAILS = oPDEXPENSE.CLAIMANT_SUFFERED_ILLNESS_DETAILS,
-                CLAIMANT_SUFFERED_ILLNESS_DATE = oPDEXPENSE.CLAIMANT_SUFFERED_ILLNESS_DATE,
-
-
-                DATE_ILLNESS_NOTICED = oPDEXPENSE.DATE_ILLNESS_NOTICED,
-                DATE_RECOVERY = oPDEXPENSE.DATE_RECOVERY,
-                DIAGNOSIS = oPDEXPENSE.DIAGNOSIS,
-                DOCTOR_NAME = oPDEXPENSE.DOCTOR_NAME,
-                DRUGS_PRESCRIBED_BOOL = oPDEXPENSE.DRUGS_PRESCRIBED_BOOL,
-                DRUGS_PRESCRIBED_DESCRIPTION = oPDEXPENSE.DRUGS_PRESCRIBED_DESCRIPTION,
-                EMPLOYEE_DEPARTMENT = oPDEXPENSE.EMPLOYEE_DEPARTMENT,
-                EMPLOYEE_NAME = oPDEXPENSE.EMPLOYEE_NAME,
-                FINANCE_APPROVAL = oPDEXPENSE.FINANCE_APPROVAL,
-                FINANCE_COMMENT = oPDEXPENSE.FINANCE_COMMENT,
-                FINANCE_NAME = oPDEXPENSE.FINANCE_NAME,
-                HOSPITAL_NAME = oPDEXPENSE.HOSPITAL_NAME,
-                HR_APPROVAL = oPDEXPENSE.HR_APPROVAL,
-                HR_COMMENT = oPDEXPENSE.HR_COMMENT,
-                HR_NAME = oPDEXPENSE.HR_NAME,
-                MANAGEMENT_APPROVAL = oPDEXPENSE.MANAGEMENT_APPROVAL,
-                MANAGEMENT_COMMENT = oPDEXPENSE.MANAGEMENT_COMMENT,
-                MANAGEMENT_NAME = oPDEXPENSE.MANAGEMENT_NAME,
-                PERIOD_CONFINEMENT_DATE_FROM = oPDEXPENSE.PERIOD_CONFINEMENT_DATE_FROM,
-                PERIOD_CONFINEMENT_DATE_TO = oPDEXPENSE.PERIOD_CONFINEMENT_DATE_TO,
-                STATUS = oPDEXPENSE.STATUS,
-                OPDTYPE = oPDEXPENSE.OPDTYPE,
-                TOTAL_AMOUNT_CLAIMED = oPDEXPENSE.TOTAL_AMOUNT_CLAIMED,
-                CLAIM_YEAR = oPDEXPENSE.CLAIM_YEAR
-            };
-
-            ViewData["OPDEXPENSE_ID"] = id;
-            return View(result2);
+                return RedirectToAction("Index", "OPDEXPENSEs");
+            }
 
         }
 
@@ -163,7 +189,7 @@ namespace OPDCLAIMFORM.Controllers
         // more details see https://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public ActionResult Edit([Bind(Include = "OPDEXPENSE_ID,EMPLOYEE_NAME,EMPLOYEE_DEPARTMENT,CLAIM_MONTH,TOTAL_AMOUNT_CLAIMED,HR_COMMENT,HR_APPROVAL,HR_APPROVAL_DATE,HR_NAME,FINANCE_COMMENT,FINANCE_APPROVAL,FINANCE_APPROVAL_DATE,FINANCE_NAME,MANAGEMENT_COMMENT,MANAGEMENT_APPROVAL,MANAGEMENT_APPROVAL_DATE,MANAGEMENT_NAME,DATE_ILLNESS_NOTICED,DATE_RECOVERY,DIAGNOSIS,CLAIMANT_SUFFERED_ILLNESS,CLAIMANT_SUFFERED_ILLNESS_DATE,CLAIMANT_SUFFERED_ILLNESS_DETAILS,HOSPITAL_NAME,DOCTOR_NAME,PERIOD_CONFINEMENT_DATE_FROM,PERIOD_CONFINEMENT_DATE_TO,DRUGS_PRESCRIBED_BOOL,DRUGS_PRESCRIBED_DESCRIPTION,OPDTYPE,STATUS,CLAIM_YEAR")] OPDEXPENSE oPDEXPENSE)
+        public ActionResult Edit([Bind(Include = "OPDEXPENSE_ID,EMPLOYEE_NAME,EMPLOYEE_DEPARTMENT,CLAIM_MONTH,TOTAL_AMOUNT_CLAIMED,HR_COMMENT,HR_APPROVAL,HR_APPROVAL_DATE,HR_NAME,FINANCE_COMMENT,FINANCE_APPROVAL,FINANCE_APPROVAL_DATE,FINANCE_NAME,MANAGEMENT_COMMENT,MANAGEMENT_APPROVAL,MANAGEMENT_APPROVAL_DATE,MANAGEMENT_NAME,DATE_ILLNESS_NOTICED,DATE_RECOVERY,DIAGNOSIS,CLAIMANT_SUFFERED_ILLNESS,CLAIMANT_SUFFERED_ILLNESS_DATE,CLAIMANT_SUFFERED_ILLNESS_DETAILS,HOSPITAL_NAME,DOCTOR_NAME,PERIOD_CONFINEMENT_DATE_FROM,PERIOD_CONFINEMENT_DATE_TO,DRUGS_PRESCRIBED_BOOL,DRUGS_PRESCRIBED_DESCRIPTION,OPDTYPE,STATUS,CLAIM_YEAR,CREATED_DATE")] OPDEXPENSE oPDEXPENSE)
         {
             if (ModelState.IsValid)
             {
@@ -179,22 +205,22 @@ namespace OPDCLAIMFORM.Controllers
         // GET: OPDEXPENSEs/Delete/5
         public ActionResult Delete(int? id)
         {
-            if (id == null)
+            if (Request.IsAuthenticated)
             {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                AuthenticateUser();
+                if (id == null)
+                {
+                    return RedirectToAction("Index", "OPDEXPENSEs");
+                }
+
+                var fileInfo = this.db.DELETE_OPDEXPENSE(id);
+
+                return RedirectToAction("Index", "OPDEXPENSEs");
             }
-
-            var fileInfo = this.db.DELETE_OPDEXPENSE(id);
-
-            return RedirectToAction("Index", "OPDEXPENSEs");
-
-
-            //OPDEXPENSE oPDEXPENSE = db.OPDEXPENSEs.Find(id);
-            //if (oPDEXPENSE == null)
-            //{
-            //    return HttpNotFound();
-            //}
-            //return View(oPDEXPENSE);
+            else
+            {
+                return RedirectToAction("Index", "OPDEXPENSEs");
+            }
         }
 
         // POST: OPDEXPENSEs/Delete/5
@@ -202,17 +228,26 @@ namespace OPDCLAIMFORM.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult DeleteConfirmed(int id)
         {
-            
-            var fileInfo = this.db.DELETE_OPDEXPENSE(id);
 
-            return RedirectToAction("Index","OPDEXPENSEs");
+            if (Request.IsAuthenticated)
+            {
+                AuthenticateUser();
+                if (id == 0)
+                {
+                    return RedirectToAction("Index", "OPDEXPENSEs");
+                }
 
+                else
+                {
+                    var fileInfo = this.db.DELETE_OPDEXPENSE(id);
+                }
 
-
-            //OPDEXPENSE oPDEXPENSE = db.OPDEXPENSEs.Find(id);
-            //db.OPDEXPENSEs.Remove(oPDEXPENSE);
-            //db.SaveChanges();
-            //return RedirectToAction("Index");
+                return RedirectToAction("Index", "OPDEXPENSEs");
+            }
+            else
+            {
+                return RedirectToAction("Index", "OPDEXPENSEs");
+            }
         }
 
         protected override void Dispose(bool disposing)
@@ -297,6 +332,14 @@ namespace OPDCLAIMFORM.Controllers
             string name = managerController.GetName();
 
             return name;
+        }
+
+        private void AuthenticateUser()
+        {
+            OFFICEAPIMANAGERController managerController = new OFFICEAPIMANAGERController();
+
+            ViewBag.RollType = managerController.AuthenticateUser();
+
         }
     }
 }
